+++ complementary/handMade.h
@@ -1,16 +1,10 @@
 //TESTS ERASE LATER
 #include <random>
 
-std::random_device rd;
-std::mt19937 rng(rd());
+extern std::random_device rd;
+extern std::mt19937 rng;
 
-float getRandomFloat(float min, float max) {
-    if(min > max){
-        std::swap(min, max);
-    }
-    std::uniform_real_distribution<float> dist(min, max);
-    return dist(rng);
-}
+float getRandomFloat(float min, float max);
 
 namespace osd{
     std::string getExecutableDir();
