+++ core/initPipelineHelper.cpp
@@ -1,4 +1,3 @@
-#pragma once
 #include "core.h"
